+++ src/main/java/br/edu/ifpb/pweb2/sorte_io/repository/ApostasRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.ifpb.pweb2.sorte_io.model.Aposta;
 
-public interface ApostasRepository extends JpaRepository<Aposta, Long> {
+public interface ApostasRepository extends JpaRepository<Aposta, Integer> {
     
 }
